Reimplemented (once more) Course.refresh as CourseRefresher.
It should build the cache files more safely now
since it works in a new directory and only updates
the database to point to it after it's successfully built.

There is also row-level locking, which should mutex other refreshes but
it might not be enough to exclude other operations. Problems ought to
be rare, so will probably deal with this much later.

Will replace Course.refresh with CourseRefresher in a subsequent commit.